Implement More Server APIs
This finishes implementing the APIs osquery needs for goquery compatibility meaning it no longer complains when using the mock server. I also have implemented the ScheduleQuery API but not a way for osquery to put results back (it just noops right now) nor a way for goquery to fetch the results (also a noop).

I also moved some code around to make it more elegant and fixed some variable that were in snake case.